Move cartesian_to_polar to .utils
